+++ docs/src/release_notes.rst
@@ -19,6 +19,8 @@ sphinx-codeautolink adheres to
   and ``highlight_language`` configuration (:issue:`166`)
 - Add :confval:`codeautolink_warn_on_default_parse_fail` to warn about
   failing to link code blocks without a language parameter (:issue:`166`)
+- Detect and implicitly use ``pycon`` lexer for blocks that look like
+  console code (:issue:`168`)
 - Fix skipping blocks with identical content to linked ones (:issue:`172`)
 
 0.16.2 (2025-01-16)